requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pytesseract==0.3.10
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
Pillow==10.4.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
orjson==3.10.7
